moved to libaligner, basics for gtest
