+++ src/aligner/similaritymatrix.cpp
