+++ src/aligner/similaritymatrix.h
